Docker and yaml added for test
