+++ .goreleaser.yaml
@@ -1,0 +1,28 @@
+project_name: gospoof
+
+builds:
+  - id: gospoof
+    main: ./cmd/gospoof
+    goos: [linux, windows, darwin]
+    goarch: [amd64, arm64]
+    ldflags: "-s -w"
+
+archives:
+  - format: tar.gz
+    name_template: "{{ .ProjectName }}_{{ .Version }}_{{ .Os }}_{{ .Arch }}"
+
+dockers:
+  - image_templates:
+      - ghcr.io/{{ .RepoOwner }}/gospoof:{{ .Version }}
+      - ghcr.io/{{ .RepoOwner }}/gospoof:latest
+    dockerfile: Dockerfile
+    build_flag_templates:
+      - --platform=linux/amd64
+
+release:
+  github:
+    owner: blackhillsinfosec
+    name: GoSpoof
+
+checksum:
+  name_template: "{{ .ProjectName }}_checksums.txt"
